requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
pyyaml>=6.0
